pyproject.toml
[tool.poetry]
name="lyceum-attendance"
version="0.1.0"
description="Telegram bot for record students attendance"
readme="README.md"
authors=["noinst"]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
aiogram="^3.26.0"
python-dotenv="^1.1.0"
sqlalchemy = "^2.0.42"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
pydantic = "^2.11.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"